requirements.txt
torch>=2.0.0
numpy>=1.23.0
pandas>=1.5.0
biopython>=1.81
scikit-learn>=1.2.0
matplotlib>=3.7.0
pyyaml>=6.0.1
tqdm>=4.66.0